+++ rust-core/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod android_jni;
+pub mod gpx;
 
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");